+++ include/effects/strip/laserline.h
@@ -35,14 +35,14 @@ extern AppTime g_AppTime;
 
 class LaserShot
 {
-    double        _position = 0.0;
-    double        _speed    = 10.0;
-    double        _size     = 10.0;
+    float         _position = 0.0;
+    float         _speed    = 10.0;
+    float         _size     = 10.0;
     uint8_t       _hue      = 0;
     
 public:
 
-    LaserShot(double position, double speed, double size, uint8_t hue)
+    LaserShot(float position, float speed, float size, uint8_t hue)
     {
         _position = position;
         _speed    = speed;
@@ -50,7 +50,7 @@ public:
         _hue      = hue;
     }
 
-    virtual bool Update(double elapsed)
+    virtual bool Update(float elapsed)
     {
         _hue += _speed * elapsed;
         _position += _speed * elapsed;
@@ -61,7 +61,7 @@ public:
 
     virtual void Draw(std::shared_ptr<GFXBase> pGFX)
     {
-        for (double d = 0; d < _size && d + _position < NUM_LEDS; d++)
+        for (float d = 0; d < _size && d + _position < NUM_LEDS; d++)
             pGFX->setPixelsF(_position + d, 1.0, CHSV(_hue + d, 255, 255), true);
     }
 };
@@ -71,12 +71,12 @@ class LaserLineEffect : public BeatEffectBase, public LEDStripEffect
   private:
     std::vector<LaserShot>      _shots;
     std::shared_ptr<GFXBase>    _gfx;
-    double                      _defaultSize;
-    double                      _defaultSpeed;
+    float                      _defaultSize;
+    float                      _defaultSpeed;
 
   public:
   
-    LaserLineEffect(double speed, double size) : 
+    LaserLineEffect(float speed, float size) : 
         BeatEffectBase(1.50, 0.00), LEDStripEffect("LaserLine"), _defaultSize(size), _defaultSpeed(speed) 
     {
     }
@@ -107,7 +107,7 @@ class LaserLineEffect : public BeatEffectBase, public LEDStripEffect
         }
     }
 
-    virtual void HandleBeat(bool bMajor, float elapsed, double span)
+    virtual void HandleBeat(bool bMajor, float elapsed, float span)
     {
         _shots.push_back(LaserShot(0.0, _defaultSpeed, _defaultSize, random8()));
     };
